KOA-5129 Ensure preview works without wrapping in composable
