+++ app/src/main/java/net/skyscanner/backpack/demo/compose/PanelStory.kt
@@ -77,6 +77,7 @@ fun DefaultPanelExample(
 }
 
 @Composable
+@Preview
 fun NoPaddingPanelExample(
   modifier: Modifier = Modifier,
 ) {
